Add product Id verification
Edit JSP formatting

+++ src/main/java/com/es/phoneshop/web/DeleteCartItemServlet.java
@@ -10,6 +10,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.text.ParseException;
 
 public class DeleteCartItemServlet extends HttpServlet {
 
@@ -23,16 +24,23 @@ public class DeleteCartItemServlet extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        String productIdString = request.getPathInfo().substring(1);
         Long productId = parseProductID(request);
-        Cart cart = cartService.getCart(request);
-
-        cartService.delete(cart, productId);
-
-        response.sendRedirect(request.getContextPath() + "/cart?modalSuccess=Cart item deleted successfully");
+        if (productId >= 0) {
+            Cart cart = cartService.getCart(request);
+            cartService.delete(cart, productId);
+            response.sendRedirect(request.getContextPath() + "/cart?modalSuccess=Cart item deleted successfully");
+        } else {
+            response.sendRedirect(request.getContextPath() + "/cart?modalError=Wrong item id");
+        }
     }
 
     private Long parseProductID(HttpServletRequest request) {
-        String productId = request.getPathInfo().substring(1);
-        return Long.valueOf(productId);
+        String productIdString = request.getPathInfo().substring(1);
+        try {
+            return Long.parseLong(productIdString);
+        } catch (NumberFormatException e) {
+            return (long) -1;
+        }
     }
 }
